requirements.txt
fastapi
openai>=1.0.0
uvicorn
python-dotenv